Refactor to Transaction domain
