+++ package.json
@@ -11,7 +11,7 @@
     "test:watch": "mocha -w",
     "test:coverage": "npm test && http-server coverage",
     "documentation": "typedoc --out documentation/ --exclude \"src/**/*.spec.ts\" --mode modules src/",
-    "documentation:serve": "npm run documentation && http-server documentation" 
+    "documentation:serve": "npm run documentation && http-server documentation"
   },
   "repository": {
     "type": "git",
