+++ src/Transaction/TransactionInput.ts
@@ -1,0 +1,12 @@
+export interface TransactionInput {
+  pointer: {
+    index: number
+    id: string
+  }
+  value: string
+  addressing?: {
+    account: number
+    change: number
+    index: number
+  }
+}
