+++ src/Transaction/TransactionOutput.ts
@@ -1,16 +1,3 @@
-export interface TransactionInput {
-  pointer: {
-    index: number
-    id: string
-  }
-  value: string
-  addressing?: {
-    account: number
-    change: number
-    index: number
-  }
-}
-
 export interface TransactionOutput {
   address: string
   value: string
