+++ src/Transaction/index.ts
@@ -1,0 +1,2 @@
+import { Transaction } from './Transaction'
+export default Transaction
