--- src/index.ts
@@ -1,2 +1,0 @@
-export default class CardanoJs {
-}
