--- src/lib/transactions/index.ts
@@ -1,1 +1,0 @@
-export { buildTransaction } from './build'
